Actualizaciones 07/06/2024

+++ Server/src/models/Cargo_Empleado.js
@@ -19,7 +19,7 @@ module.exports = (sequelize) => {
       allowNull: false,
     },
     salario: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(11, 2),
       allowNull: false,
     },
     fecha_ingreso: {
